Implemented pratt parsing

+++ src/calc/common.rs
@@ -1,2 +1,2 @@
 pub type Integer = u128;
-pub type Float = u64;
+pub type Float = f32;
